Gracefully handle vanishing /proc/[pid] directories
This is commonly encountered for short lived processes; we see the
process when initially crawling the proc filesystem, but then don't
find it when we attempt to parse each discovered status file. In order
to prevent erroring out over short-lived files we skip to the next
file in the list.

Applied light refactoring to further simplify main func. Will
likely rename the function later.

refs GH-19